requirements.txt
strands-agents 
strands-agents[openai]
strands-agents-tools
notebook
PyPDF2